requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database clients
supabase>=2.0.0
neo4j>=5.0.0

# AI/ML
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.0.0
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
